chore(middleware-retry): move non-config interfaces to types.ts
