+++ packages/middleware-retry/src/defaultRetryQuota.ts
@@ -1,7 +1,7 @@
 import { SdkError } from "@aws-sdk/smithy-client";
 
 import { NO_RETRY_INCREMENT, RETRY_COST, TIMEOUT_RETRY_COST } from "./constants";
-import { RetryQuota } from "./defaultStrategy";
+import { RetryQuota } from "./types";
 
 export const getDefaultRetryQuota = (initialRetryTokens: number): RetryQuota => {
   const MAX_CAPACITY = initialRetryTokens;
